requirements.txt
2to3==1.0
anyio==3.6.2
beautifulsoup4==4.12.0
binwalk==2.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
decorator==5.1.1
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
MouseInfo==0.1.3
pyarmor==7.7.4
pyarmor.cli.core==2.1.5
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.17
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==4.10.0
python-telegram-bot==12.8
pytweening==1.0.4
requests==2.28.2
rfc3986==1.5.0
semantic-version==2.10.0
setuptools-rust==1.5.2
sniffio==1.3.0
soupsieve==2.4
stem==1.8.1
tornado==6.2
typing_extensions==4.5.0
urllib3==1.26.15
